// src/models/Property.ts
import mongoose, { Schema, Document } from 'mongoose';

export interface IProperty extends Document {
  name: string;
  city: string;
  area: string;
  address: string;
  description?: string;
  photos: string[];
  ownerId: mongoose.Types.ObjectId;
  isActive: boolean;
  rating?: number;
  genderAllowed: 'any' | 'male' | 'female';
  isVerified: boolean;
  priceRange?: string;
  createdAt: Date;
  updatedAt: Date;
}

const PropertySchema: Schema = new Schema(
  {
    name: { type: String, required: true },
    city: { type: String, required: true },
    area: { type: String, required: true },
    address: { type: String, required: true },
    description: { type: String },
    photos: [{ type: String }],
    ownerId: { type: Schema.Types.ObjectId, ref: 'User', required: true },
    isActive: { type: Boolean, default: true },
    rating: { type: Number, default: 0 },
    genderAllowed: { type: String, enum: ['any', 'male', 'female'], default: 'any' },
    isVerified: { type: Boolean, default: false },
    priceRange: { type: String },
  },
  { timestamps: true }
);

export default mongoose.models.Property || mongoose.model<IProperty>('Property', PropertySchema);